Support zigbee lights and the temperature sensor of the Echo device.
This moves all of the phoenix calls being done by the alarm platform into the date update coordinator so that sensor/light can share. This involved significant refactoring of the guard code.

Device support intentionally excludes all Skill-based devices Alexa has access to as there is likely another Home Assistant integration that can better address those items. The light support still remains quite limited supporting only brightness and power state. Temperature sensor support is complete, but is restricted specifically to the sensor built-in to an Echo.

With the exception of the new guard code, everything should be hidden by an new option. Enabling this feature requires turning that option and reloading the integration.