fix: fix issue where next_alarm would cancel on reload (#1379)
Where multiple instances of the component are in use, the "dev_id" is made up
from a combination of the serial number & account email address.

This was being used when trying to load out the notifications from the
async_update call. The key that should have been used is only the serial
number of the device. This commit changes to using the serial.