fix: store and reuse mac_dms between sessions
Store mac_dms in the config. This should reduce the frequency of reauth
requirements if registration is failing. This does not remove the requirement for
Amazon to provide a valid mac_dms token at least once. Users who are
unable to receive a valid mac_dms token will need to use the unsupported
3.x series.

closes #1620