fix: speaker group status indication
This fixes issue #2993.

Fixes speaker group status display.

The speaker group status display is no longer displayed because the
"clusterMembers" and "parentClusters" data are no longer available from
the API. As a workaround, "clusterMembers" is now created from
"lemurVolume.memberVolume". However, there are the following
limitations:

1. For stereo speakers, only the parent entity will be updated.
2. If a speaker group contains stereo speakers, the parent entity of the
stereo configuration will not be updated. Only the individual speakers
in the stereo configuration will be updated.